ItemStack is o more. Now count and NBT is directly on the items. Seems to work at first glance, but this was a rather large refactoring so expect bugs.

+++ src/MiNET/MiNET/Items/FoodItem.cs
@@ -6,7 +6,7 @@ namespace MiNET.Items
 	{
 		public int FoodPoints { get; set; }
 
-		public FoodItem(int id, short metadata, int foodPoints) : base(id, metadata)
+		public FoodItem(short id, short metadata, int foodPoints) : base(id, metadata)
 		{
 			FoodPoints = foodPoints;
 		}
